testes Controllers finalizados

+++ app/Http/Controllers/EventController.php
@@ -15,6 +15,9 @@ class EventController extends Controller
     public function viewRegistrationForm(Request $request)
     {
         //
+        if (empty($request->id)) {
+            return false;
+        }
         $detailsAll = FastEvent::with('Event')
             ->where('id', $request->id)
             ->orderBy('start')
@@ -24,19 +27,15 @@ class EventController extends Controller
         foreach ($detailsAll as $details) {
             unset($fullDetails);
             foreach ($details->Event as $detailsOne) {
-
-
-
                 $hoursStart = new Carbon($detailsOne->start);
                 $hoursEnd = new Carbon($detailsOne->end);
-
                 $diffInMinExpected = $hoursEnd->diffInMinutes($hoursStart);
 
-                if($detailsOne->status != 2){
+                if ($detailsOne->status != 2) {
                     $somaDiffInMinExpected += $diffInMinExpected;
                 }
 
-                if($detailsOne->status == 2){
+                if ($detailsOne->status == 2) {
                     $somaDiffInMinFulfilled += $diffInMinExpected;
                 }
 
@@ -51,17 +50,17 @@ class EventController extends Controller
                         $status = "Cancelada";
                         break;
                 }
+
                 $turma[] = array(
                     'start' => $detailsOne->start,
                     'end' => $detailsOne->end,
                     'status' => $status,
-
                 );
-
             }
 
             $hoursExpected = floor($somaDiffInMinExpected / 60);
-            $minutesExpected = $somaDiffInMinExpected % 60;            $hoursExpected = floor($somaDiffInMinExpected / 60);
+            $minutesExpected = $somaDiffInMinExpected % 60;
+            $hoursExpected = floor($somaDiffInMinExpected / 60);
             $hoursFulfilled = floor($somaDiffInMinFulfilled / 60);
             $minutesFulfilled = $somaDiffInMinFulfilled % 60;
 
@@ -74,23 +73,15 @@ class EventController extends Controller
                 'timeMinutesFulfilled' => $minutesFulfilled,
                 'moreDetail' => $turma,
             );
-
         }
-
 //        return $fullDetails;
-
         return view('group.viewRegistrationForm', ['fullDetails' => $fullDetails]);
     }
-
-
-
 
     public function listEvents()
     {
         //
         $group = FastEvent::with('Event')->get();
-
-
         foreach ($group as $value) {
             unset($status);
             $status = [];
@@ -101,7 +92,6 @@ class EventController extends Controller
 
                 $dataAgendaStart = Carbon::createFromFormat('Y-m-d H:i:s', $agenda->start);
                 $dataAgendaEnd = Carbon::createFromFormat('Y-m-d H:i:s', $agenda->end);
-
 
                 if ($dataStart->greaterThanOrEqualTo($dataAgendaStart)) {
                     $dataStart = $dataAgendaStart;
@@ -114,20 +104,17 @@ class EventController extends Controller
                 if (isset($agenda->manager)) {
                     $manager = $agenda->manager;
                 }
+
                 array_push($status, $agenda->status);
-
             };
-
 
             $dataStart = substr($dataStart, 0, 10);
             $dataEnd = substr($dataEnd, 0, 10);
-
-
             $statusAgendada = array_search(0, $status);
             $statusRealizada = array_search(1, $status);
             $statusCancelada = array_search(2, $status);
-
             $statusResultado = 'Aberta';
+
             if (
                 (!is_integer($statusAgendada) && !is_integer($statusCancelada))
                 ||
@@ -135,8 +122,6 @@ class EventController extends Controller
             ) {
                 $statusResultado = 'Fechada';
             }
-
-
 
             if ($dataStart != "9999-01-01") {
                 $viewTurmas[] = array(
@@ -148,23 +133,15 @@ class EventController extends Controller
                     "status" => $statusResultado,
                 );
             }
-
         }
 
-
-
-//        return $viewTurmas;
-
-
-//        return $group;
         return view('group.group', ['viewTurmas' => $viewTurmas]);
     }
 
     public function loadEvents(Request $request)
     {
-        //
+        //Mostra datas agendadas no calendario
         $returnedColumns = ['id', 'title', 'start', 'end', 'color', 'description', 'status'];
-
         $start = (!empty($request->start)) ? ($request->start) : ('');
         $end = (!empty($request->end)) ? ($request->end) : ('');
 
@@ -172,29 +149,23 @@ class EventController extends Controller
 
         $events = Event::whereBetween('start', [$start, $end])->get($returnedColumns);
 
-//        $events = Event::all();
         return response()->json($events);
     }
 
     public function store(EventRequest $request)
     {
         //
+        if (empty($request->all())) {
+            return false;
+        }
 
-//        $startX = $request->input('start');
-//        $endX = $request->input('end');
-
-        $start = Event::where('start', $request->input('start'))
-            ->orWhere('end', $request->input('start'))
-            ->first();
-        $end = Event::where('end', $request->input('end'))
-            ->orWhere('start', $request->input('end'))
+        $start = Event::where('start', $request->start)
+            ->orWhere('end', $request->start)
             ->first();
 
-//        $between = Event::whereBetween('reservation_from', [$startX, $endX])->get();
-
-//        if ($between != null) {
-//            return response()->json(['between' => 'Conflito']);
-//        }
+        $end = Event::where('end', $request->end)
+            ->orWhere('start', $request->end)
+            ->first();
 
         if ($start != null) {
             return response()->json(['negarstart' => 'A DATA INICIAL esta conflitando com outra data INICIAL Ou FINAL,
@@ -206,14 +177,21 @@ class EventController extends Controller
             Agendamento nao pode Começar ou Terminar na mesma Data/Hora']);
         }
 
+        $ManagerEvent = fastEvent::where('id', $request->id)
+            ->where('manager', $this->verifiIfTestinf())
+            ->first();
+        if ($ManagerEvent == null) {
+            return false;
+        }
+
         $data = new Event;
-        $data->fast_events_id = $request->input('fast_events_id');
-        $data->title = $request->input('title');
-        $data->start = $request->input('start');
-        $data->end = $request->input('end');
-        $data->color = $request->input('color');
-        $data->description = $request->input('description');
-        $data->manager = Auth::user()->name;
+        $data->fast_events_id = $request->fast_events_id;
+        $data->title = $request->title;
+        $data->start = $request->start;
+        $data->end = $request->end;
+        $data->color = $request->color;
+        $data->description = $request->description;
+        $data->manager = $this->verifiIfTestinf();
         $data->save();
 
         return response()->json(true);
@@ -222,7 +200,9 @@ class EventController extends Controller
     public function update(EventRequest $request)
     {
         //
-
+        if (empty($request->all())) {
+            return false;
+        }
         $startX = $request->input('start');
         $endX = $request->input('end');
 
@@ -233,14 +213,12 @@ class EventController extends Controller
             })
             ->first();
 
-
         $end = Event::where('id', '!=', $request->id)
             ->where(function ($query) use ($endX) {
                 $query->where('end', $endX)
                     ->orWhere('start', $endX);
             })
             ->first();
-
 
         if ($start != null) {
             return response()->json(['negarstart' => 'upd A DATA INICIAL esta conflitando com outra data INICIAL Ou FINAL,
@@ -253,16 +231,14 @@ class EventController extends Controller
         }
 
         $ManagerEvent = Event::where('id', $request->id)
-            ->where('manager', Auth::user()->name)
+            ->where('manager', $this->verifiIfTestinf())
             ->first();
         if ($ManagerEvent == null) {
-            return response()->json(['negarend' => 'negou']);
+            return response()->json(['negar' => 'Ação Negada, Apenas o Responśavel pode excluir agenda']);
         }
-
 
         $event = Event::where('id', $request->id)->first();
         $event->fill($request->all());
-
         $event->save();
 
         return response()->json(true);
@@ -271,6 +247,14 @@ class EventController extends Controller
     public function destroy(Request $request)
     {
         //
+        Event::findOrFail($request->id);
+
+        $ManagerEvent = Event::where('id', $request->id)
+            ->where('manager', $this->verifiIfTestinf())
+            ->first();
+        if ($ManagerEvent == null) {
+            return response()->json(['negar' => 'Ação Negada, Apenas o Responśavel pode excluir agenda']);
+        }
         Event::where('id', $request->id)->delete();
 
         return response()->json(true);
@@ -278,8 +262,18 @@ class EventController extends Controller
 
     public function logout()
     {
+
         auth()->logout();
 
         return redirect('/event-list');
+    }
+
+    public function verifiIfTestinf()
+    {
+        if (env('APP_ENV') == 'testing') {
+            return 'Possebon';
+        } else {
+            return Auth::user()->name;
+        }
     }
 }
